backend/requirements.txt
flask==2.0.1
pymongo==3.12.0
flask-cors==3.0.10
pyjwt==2.1.0
passlib==1.7.4
python-dotenv==0.19.0 